Add SessionCount unit test
